requirements.txt
Django==1.11.3
Pygments==2.2.0
argparse==1.2.1
certifi==2017.4.17
chardet==3.0.4
djangorestframework==3.6.3
drf-generators==0.3.0
httpie==0.9.9
idna==2.5
mysqlclient==1.3.10
pytz==2017.2
requests==2.18.1
urllib3==1.21.1
wsgiref==0.1.2
